+++ RustyBags/Plugin.cs
@@ -11,12 +11,14 @@ using UnityEngine;
 
 namespace RustyBags
 {
+    [BepInDependency("Azumatt.AzuCraftyBoxes", BepInDependency.DependencyFlags.SoftDependency)]
+    [BepInDependency("ZenDragon.ZenConstruction", BepInDependency.DependencyFlags.SoftDependency)]
     [BepInDependency("Azumatt.AzuExtendedPlayerInventory", BepInDependency.DependencyFlags.SoftDependency)]
     [BepInPlugin(ModGUID, ModName, ModVersion)]
     public class RustyBagsPlugin : BaseUnityPlugin
     {
         internal const string ModName = "RustyBags";
-        internal const string ModVersion = "1.1.2";
+        internal const string ModVersion = "1.1.3";
         internal const string Author = "RustyMods";
         private const string ModGUID = Author + "." + ModName;
         public static readonly string ConfigFileName = ModGUID + ".cfg";
@@ -246,15 +248,12 @@ namespace RustyBags
                     });
             }
             
+            BagCraft.Init();
+            
             Assembly assembly = Assembly.GetExecutingAssembly();
             _harmony.PatchAll(assembly);
         }
-
-        public void Start()
-        {
-            BagCraft.CheckOtherPatches();
-        }
-
+        
         private void OnDestroy()
         {
             Config.Save();